+++ src/main/java/com/company/financial/app/domain/model/dto/ClientDto.java
@@ -19,6 +19,6 @@ public class ClientDto {
     private Long birthdate;
     private Long creationDate;
     private Long modificationDate;
-    private double moneyInvested;
+    private int moneyInvested;
     private double percentageInvested;
 }
