+++ src/main/java/com/company/financial/app/domain/model/dto/request/ClientRequest.java
@@ -16,6 +16,6 @@ public class ClientRequest implements Serializable {
     private String lastName;
     private String email;
     private Long birthdate;
-    private double moneyInvested;
+    private int moneyInvested;
     private String operatorComparator;
 }
